# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Medcom API",
    "description": "Medcom API documentation",
    "termsOfService": "",
    "contact": {
      "name": "",
      "url": "",
      "email": ""
    },
    "license": {
      "name": "unlicensed",
      "url": ""
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/notifications/{id}/mark-as-read": {
      "put": {
        "tags": ["notification-resource"],
        "operationId": "markNotificationAsRead",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": ["notification-resource"],
        "operationId": "getAllNotifications",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "eagerload",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/{id}": {
      "get": {
        "tags": ["notification-resource"],
        "operationId": "getNotification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["notification-resource"],
        "operationId": "deleteNotification",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/notifications/unread": {
      "get": {
        "tags": ["notification-resource"],
        "operationId": "getUnreadNotifications",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications/my": {
      "get": {
        "tags": ["notification-resource"],
        "operationId": "getMyNotifications",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/my-pharmacy": {
      "get": {
        "tags": ["notification-resource"],
        "operationId": "getMyPharmacyNotifications",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PharmacyDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "ProductDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "officialUrl": {
            "type": "string"
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "login": {
            "type": "string"
          }
        }
      },
      "WatchListItemDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "lastAvailability": {
            "type": "boolean"
          },
          "lastAvailabilityTime": {
            "type": "string",
            "format": "date-time"
          },
          "notifyAllUsers": {
            "type": "boolean"
          },
          "createdBy": {
            "$ref": "#/components/schemas/UserDTO",
            "readOnly": true
          },
          "product": {
            "$ref": "#/components/schemas/ProductDTO"
          },
          "pharmacy": {
            "$ref": "#/components/schemas/PharmacyDTO",
            "readOnly": true
          }
        }
      },
      "NotificationDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "notificationType": {
            "type": "string",
            "enum": ["SMS", "EMAIL", "WHATSAPP"]
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "readAt": {
            "type": "string",
            "format": "date-time"
          },
          "sent": {
            "type": "boolean"
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          },
          "delivered": {
            "type": "boolean"
          },
          "deliveredAt": {
            "type": "string",
            "format": "date-time"
          },
          "failed": {
            "type": "boolean"
          },
          "failedAt": {
            "type": "string",
            "format": "date-time"
          },
          "failureReason": {
            "type": "string"
          },
          "externalMessageId": {
            "type": "string"
          },
          "pharmacy": {
            "$ref": "#/components/schemas/PharmacyDTO",
            "readOnly": true
          },
          "user": {
            "$ref": "#/components/schemas/UserDTO",
            "readOnly": true
          },
          "watchListItem": {
            "$ref": "#/components/schemas/WatchListItemDTO"
          }
        }
      },
...
    }
  }
}